Updated test comparing diagonal of cholSigmaInv
cholSigmaInv in the WLS case (heterogeneous variances, but not mixed
model) is a vector - the diagonal of the cholSigmaInv matrix returned
by a mixed model function when all variance components corresponding to
between-individuals correlations are equal to zero. Here we fixed the
test function to compare the vector from WLS function to the diagonal
from the mixed model function.